+++ src/main/kotlin/nextstep/mission/LottoApplication.kt
@@ -1,15 +1,16 @@
 package nextstep.mission
 
-import nextstep.mission.lotto.Lottos
-import nextstep.mission.lotto.WinningResult
+import nextstep.mission.lotto.Lotto
+import nextstep.mission.lotto.LottoMachine
 import nextstep.mission.lotto.io.ConsoleInput
 import nextstep.mission.lotto.io.ConsoleOutput
+import nextstep.mission.lotto.vo.LottoNumbers
+import nextstep.mission.lotto.vo.WinningResult
 
 fun main() {
     val inputPrice: Int = ConsoleInput.inputPrice()
-    val lottos = Lottos(inputPrice)
-    ConsoleOutput.printLottos(lottos.lottos)
-    val winningNumbers: List<Int> = ConsoleInput.inputWinningNumbers()
-    val winningResult: WinningResult = lottos.checkWinningNumbers(winningNumbers)
-    ConsoleOutput.printWinningResult(winningResult, winningResult.rateOfReturn(inputPrice))
+    val lotto: Lotto = LottoMachine.create(inputPrice)
+    ConsoleOutput.printLotto(lotto)
+    val winningNumbers: LottoNumbers = ConsoleInput.inputWinningNumbers2()
+    val result: WinningResult = lotto.matchWinningNumbers(winningNumbers)
 }
