+++ src/main/kotlin/nextstep/mission/lotto/vo/WinningResult.kt
@@ -3,4 +3,4 @@ package nextstep.mission.lotto.vo
 import nextstep.mission.lotto.WinningPrize
 
 @JvmInline
-value class WinningResult2(private val result: Map<WinningPrize, Int>)
+value class WinningResult(private val result: Map<WinningPrize, Int>)
