+++ src/test/kotlin/nextstep/mission/lotto/LottoMachineTest.kt
@@ -8,7 +8,7 @@ class LottoMachineTest : StringSpec({
     "로또 기계는 입력받은 금액만큼의 로또 숫자 리스트를 자동 생성한다." {
         val totalLottoPrice = 10_000
 
-        val result: Lotto2 = LottoMachine.create(totalLottoPrice)
+        val result: Lotto = LottoMachine.create(totalLottoPrice)
 
         result.lottoNumbers.size shouldBe 10
     }
